+++ src/index.js
@@ -4,10 +4,11 @@ import {
   addTask,
   deleteTaks,
   updateTaks,
-  changeStatus,
+  deleteAllTaks,
 } from './modules/functions.js';
+import changeStatus from './modules/statusUpdate.js';
 
-const ul = document.querySelector('ul');
+const ul = document.querySelector('.ul');
 loadTasks(localStorage.getItem('lists'), ul);
 const addInput = document.querySelector('.add-input form');
 addInput.addEventListener('submit', (event) => {
@@ -15,8 +16,12 @@ addInput.addEventListener('submit', (event) => {
   addTask(document.querySelector('#new-item'), ul);
   deleteTaks(ul);
   updateTaks();
-  changeStatus();
+  changeStatus(ul);
 });
 deleteTaks(ul);
 updateTaks();
 changeStatus();
+const clearAll = document.querySelector('.btn-clear button');
+clearAll.addEventListener('click', () => {
+  deleteAllTaks(ul);
+});
